% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_col}
\alias{ca_col}
\title{Get a single church army colour by name}
\usage{
ca_col(x)
}
\arguments{
\item{x}{A character. Value must be one of \link[carutools]{ca_sample_cols}}
}
\description{
A wrapper around \link"[carutools]{ca_cols} that strictly accepts one name and returns one colour
}
