% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_colours.R
\name{ca_colour}
\alias{ca_colour}
\title{Get single church army colour}
\usage{
ca_colour(x)
}
\arguments{
\item{x}{A string - must be one of \link[carutools]{ca_sample_cols}}
}
\value{
A single colour hexcode
}
\description{
Get single church army colour
}
