% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_colours.R
\name{ca_colours}
\alias{ca_colours}
\title{Get church army colours by name}
\usage{
ca_colours(...)
}
\arguments{
\item{...}{Any number of characters vectors. All values must all be in \link[carutools]{ca_sample_cols}}
}
\value{
A character vector
}
\description{
Get church army colours by name
}
\examples{
ca_colours("cyan", "green", "purple")
ca_colour("darkteal")
## colour names are forgiving:
ca_colour("dark-teal")
}
