% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_black}
\alias{ca_black}
\alias{ca_white}
\alias{ca_orange}
\alias{ca_gold}
\alias{ca_green}
\alias{ca_maroon}
\alias{ca_cyan}
\alias{ca_purple}
\alias{ca_light_teal}
\alias{ca_dark_teal}
\alias{ca_brown}
\title{Church Army Colours}
\usage{
ca_black()

ca_white()

ca_orange()

ca_gold()

ca_green()

ca_maroon()

ca_cyan()

ca_purple()

ca_light_teal()

ca_dark_teal()

ca_brown()
}
\value{
A single character vector of length one.
}
\description{
These functions each return the hexadecimal code pertaining to a colour in Church Army's brand
}
\details{
These functions are wrappers around character vectors of length one. They take no arguments and always return the same string.
}
