% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_pal}
\alias{ca_pal}
\title{Get a church army palette by name}
\usage{
ca_pal(which_pal = NULL)
}
\arguments{
\item{which_pal}{A character, which must be one of \link[carutools]{ca_sample_pals}.}
}
\value{
A character vector of length 5
}
\description{
This function returns a Church Army palette specified by the user
}
\details{
Depending on which of the names in \code{get_pal()} is passed to \code{which_pal},
the hexcodes of one of Church Army's colour palettes is returned

\code{which_pal} must exactly match one of \code{get_pal()}, otherwise an error is thrown.
}
