% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_pal_orange}
\alias{ca_pal_orange}
\alias{ca_pal_gold}
\alias{ca_pal_light_teal}
\alias{ca_pal_cyan}
\alias{ca_pal_dark_teal}
\alias{ca_pal_maroon}
\alias{ca_pal_purple}
\alias{ca_pal_brown}
\alias{ca_pal_green}
\alias{ca_pal_mix}
\title{Church Army Colour Pallets}
\usage{
ca_pal_orange()

ca_pal_gold()

ca_pal_light_teal()

ca_pal_cyan()

ca_pal_dark_teal()

ca_pal_maroon()

ca_pal_purple()

ca_pal_brown()

ca_pal_green()

ca_pal_mix()
}
\value{
A single character vector of length five.
}
\description{
These functions each return a colour pallet from Church Army's brand
}
\details{
These functions are wrappers around character vectors of length five.
They take no arguments and always return the same character vector.
}
