% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_sample_pals}
\alias{ca_sample_pals}
\alias{ca_sample_cols}
\title{See all Church Army palette and colour names}
\usage{
ca_sample_pals()

ca_sample_cols()
}
\value{
A character vector of Church Army palette or colour names
}
\description{
A character vector of colour or pallette names accepted by \code{carutools} functions. For instance,
\link[carutools]{ca_pal} will only accept a name from \code{ca_sample_pals()}.
}
