% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{ca_scale_colour_continuous}
\alias{ca_scale_colour_continuous}
\alias{ca_scale_fill_continuous}
\title{Continuous colour/fill scales for ggplot2}
\usage{
ca_scale_colour_continuous(ca_palette = "orange", ...)

ca_scale_fill_continuous(ca_palette, ...)
}
\arguments{
\item{ca_palette}{The CA palette to use for the scale. Must be one of \link[carutools]{ca_sample_pals}}

\item{...}{Other arguments passed to \link[ggplot2]{scale_colour_gradientn}/\link[ggplot2]{scale_fill_discrete}}
}
\description{
Add Church Army colour/fill scales to ggplot objects
}
\details{
These functions are wrappers around \link[ggplot2]{scale_colour_gradientn} that add Church Army colours to the aesthetics of ggplot objects.
Note that, since CA is a UK/ROI based charity, there are no americanised aliases for 'colour'.

These functions take a single argument \code{ca_palette}, which is passed to \link[carutools]{ca_pal} to specify a palette.
The values of the palette are used to generate a continuous scale.
}
