% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{ca_scale_colour_discrete}
\alias{ca_scale_colour_discrete}
\alias{ca_scale_fill_discrete}
\title{Discrete colour/fill scales for ggplot2}
\usage{
ca_scale_colour_discrete(ca_palette = "mix", ...)

ca_scale_fill_discrete(ca_palette = "mix", ...)
}
\arguments{
\item{ca_palette}{The CA palette to use for the scale. A character vector. If length one, it must be one of \link[carutools]{ca_sample_pal}.
Otherwise, every value must be one of \link[carutools]{ca_cols}}

\item{...}{Other arguments passed to \link[ggplot2]{scale_colour_discrete}/\link[ggplot2]{scale_fill_discrete}}
}
\description{
Add Church Army colour/fill scales to ggplot objects
}
\details{
These functions are wrappers around \link[ggplot2]{scale_colour_discrete} that add Church Army colours to the aesthetics of ggplot objects.
Note that, since CA is a UK/ROI based charity, there are no americanised aliases for 'colour'.

These functions take a single argument \code{ca_palette}. If its length is one, it is passed to \link[carutools]{ca_pal} to specify a palette.
Otherwise, it is passed to \link[carutools]{ca_cols} to specify individual colours.`
}
