% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{ca_scale_colour_gradient}
\alias{ca_scale_colour_gradient}
\alias{ca_scale_fill_gradient}
\alias{ca_scale_colour_gradient2}
\alias{ca_scale_fill_gradient2}
\title{Continuous colour/fill scales for ggplot2}
\usage{
ca_scale_colour_gradient(low = "cyan", high = "green", ...)

ca_scale_fill_gradient(low = "cyan", high = "green", ...)

ca_scale_colour_gradient2(low = "cyan", high = "green", mid = "white", ...)

ca_scale_fill_gradient2(low = "cyan", high = "green", mid = "white", ...)
}
\arguments{
\item{...}{Other arguments passed to \link[ggplot2]{scale_colour_gradientn}/\link[ggplot2]{scale_fill_discrete}}
}
\description{
Create continuous colour gradients from Church Army colours, and use these as scales for ggplot objects
}
\details{
These functions are wrappers around \link[ggplot2]{scale_colour_gradient}/\link[ggplot2]{scale_colour_gradient2} that add Church Army colours to the aesthetics of ggplot objects.
Note that, since CA is a UK/ROI based charity, there are no americanised aliases for 'colour'.

Arguments for \code{low}, \code{high} and \code{mid} are passed to  \link[carutools]{ca_col} to specify a Church Army colour that is forwarded to the gradient-making function.
The values of the palette are used to generate a continuous scale.
}
