% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-scales.R
\name{ca_scale_fill_brew}
\alias{ca_scale_fill_brew}
\alias{ca_scale_colour_brew}
\title{Church Army Discrete colour/fill Scales, brewed from palettes}
\usage{
ca_scale_fill_brew(low = "cyan", high = "green", mid = ca_white(), ...)

ca_scale_colour_brew(low = "cyan", high = "green", mid = ca_white(), ...)
}
\arguments{
\item{low}{The CA palette to be used for the lower half of the scale. Must be one of \code{ca_show_pal()}.}

\item{high}{The CA palette to be used for the upper half of the scale. Must be one of \code{ca_show_pal()}.}

\item{mid}{The colour to use for the mid-point of the scale. Must be a colour that ggplot can understand.
If this is null, a new colour is calculated by blending the mid-points of low and high scales.}
}
\description{
Discrete, diverging scales with a white midpoint, useful for plotting likert responses
}
\details{
These are wrappers around \link[ggplot2]{scale_colour_discrete}/\link[ggplot2]{scale_fill_discrete}.
'Low' and 'High' arguments are passed to \link[carutools]{ca_pal} to get official Church Army
palettes, from which values are extracted to be used in the discrete scale.

By default, odd scales have a \code{ca_white()} mid-point, which is convenient for plotting likert responses.
Specifying \code{mid = NULL} replaced this with an automatically mixed colour, which often produces
an attractive diverging scale. With the exception of this mixed colour, all colours in the scale
will be official CA brand colours
}
