% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%+na\%}
\alias{\%+na\%}
\title{Add vectors, ignoring NAs}
\usage{
x \%+na\% y
}
\value{
A numeric vector
}
\description{
Vectorised addition with built-in missingness handling. If both elements are missing, \code{NA} is returned. Otherwise missing values are treated as zero. A \code{+} (i.e vectorised) equivalent to \code{sum(..., na.rm = TRUE)}. Vectors are recycled when lengths do not match.
}
\examples{
c(1, 2, NA, 4, NA) \%+na\% c(1, 2, 3, NA, NA)
}
