% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{oranges}
\alias{oranges}
\title{Get a single church army colour by name}
\format{
An object of class \code{character} of length 5.
}
\usage{
oranges
}
\description{
Get a single church army colour by name
}
\keyword{datasets}
