% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{range_labels}
\alias{range_labels}
\title{Range labels}
\usage{
range_labels(
  x,
  width,
  sep = " - ",
  explicit_zero = TRUE,
  suffix = "",
  continuous = NULL,
  ordered = TRUE
)
}
\arguments{
\item{x}{a numeric vector}

\item{width}{Desidered label width - numeric}

\item{sep}{The character used to separate values within each label}

\item{explicit_zero}{Logical; should zeroes have a label all to themselves? They do by default.}

\item{suffix}{Optional suffix to append to each label (e.g. "\%")}

\item{continuous}{If NULL, labels assume integer values. Supply a value here to determine the precision with which a continuous range is separated.}

\item{ordered}{Logical; should output be an ordered factor?}
}
\value{
A character vector of labels
}
\description{
Create 'range labels' for grouping and tabulating a continuous variable
}
\examples{
range_labels(1:10, width = 3)
range_labels(1:10/2, width = 1, continuous = 0.1)
}
