% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeller.R
\name{relabeller}
\alias{relabeller}
\title{Relabeller}
\usage{
relabeller(...)
}
\value{
A relabelling function
}
\description{
A function factory that constructs a \link[dplyr]{case_match} function. Useful for the \code{labels} argument of ggplot's \verb{scale_*_*} functions.
}
\examples{
fruits <- c("apple", "pear", "pear", "banana")

fruit_relabel <- relabeller("pear" ~ "orange", "banana" ~ "plum")
fruit_relabel(fruits)

## Normal use case
\dontrun{
ggplot(palmerpenguins::penguins, aes(x = sex, y = flipper_length_mm)) +
  geom_point(alpha = 0.5) +
  scale_x_discrete(labels = relabeller("male" ~ "BOY\nPENGUIN", "female" ~ "GIRL\nPENGUIN"))
}
}
