% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally-delimited-string.R
\name{tally_delimited_string}
\alias{tally_delimited_string}
\title{Tally delimited string}
\usage{
tally_delimited_string(x, col)
}
\arguments{
\item{x}{A data-frame, tibble or similar}

\item{col}{\if{html}{\out{<dynamic>}} The character column to tally}
}
\value{
A data-frame-like object of the same type as \code{x}
}
\description{
Expand a delimited string column into columns that tally each instance of every observed delimited value.
}
\examples{
df <- data.frame(name = c("anna", "betty"),
              fruits = c("apple, banana", "pear, banana"))

tally_delimited_string(df, fruits)

}
