% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ca.R
\name{theme_ca}
\alias{theme_ca}
\title{Church Army theme for ggplot2 visualisations}
\usage{
theme_ca(colour = "cyan", ...)
}
\arguments{
\item{colour}{A string, one of \code{\link[=ca_sample_pals]{ca_sample_pals()}}}

\item{...}{Additional arguments passed on to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
A ggplot2 theme object
}
\description{
Church Army theme for ggplot2 visualisations
}
