% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{ca_sample_pals}
\alias{ca_sample_pals}
\title{See all Church Army palette names}
\usage{
ca_sample_pals()
}
\value{
A character vector of nine Church Army pallette names
}
\description{
A character vector of pallette names accepted by \code{carutools} functions. For instance,
\link[carutools]{ca_pal} will only accept a name from these characters as its output.
}
