% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_labels.R
\name{range_labels}
\alias{range_labels}
\title{Range labels}
\usage{
range_labels(
  x,
  width,
  start = NA,
  sep = " - ",
  explicit_zero = TRUE,
  suffix = "",
  include = c("upper", "lower", "integer"),
  ordered = TRUE
)
}
\arguments{
\item{x}{a numeric vector}

\item{width}{Desidered label width - numeric}

\item{start}{The number from which to start the lowest range label}

\item{sep}{The character used to separate values within each label}

\item{explicit_zero}{Logical; should zeroes have a label all to themselves? They do by default.}

\item{suffix}{Optional suffix to append to each label (e.g. "\%")}

\item{include}{Which side of each range should be included. One of 'upper', 'lower' or 'integer' - the latter is a special case that deals exclusively with integer ranges and otherwise crashes.}

\item{ordered}{Logical; should output be an ordered factor?}
}
\value{
A character vector of labels
}
\description{
Create 'range labels' for grouping and tabulating a continuous variable
}
\examples{
range_labels(1:10, width = 3)
range_labels(1:10, width = 3, include = "lower", start = 0, explicit_zero = TRUE)
}
