% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally-delimited-string.R
\name{tally_delimited_string}
\alias{tally_delimited_string}
\title{Tally delimited string}
\usage{
tally_delimited_string(
  x,
  col,
  count = FALSE,
  names_repair = carutools:::repair_names,
  names_prefix = NULL
)
}
\arguments{
\item{x}{A data-frame, tibble or similar}

\item{col}{\if{html}{\out{<dynamic>}} The character column to tally}

\item{count}{TRUE/FALSE. Should items in strings be counted or just marked as present/missing? These options respectively result in new columns being integer or logical type.}

\item{names_repair}{Function to repair names of new columns, prior to the appending of prefix}

\item{names_prefix}{Prefix for names of new columns. If NULL (the default), the name of \code{col} is used.}
}
\value{
A data-frame-like object of the same type as \code{x}
}
\description{
Expand a delimited string column into columns that tally each instance of every observed delimited value.
}
\examples{
df <- data.frame(name = c("anna", "betty"),
              fruits = c("apple, banana", "pear, banana, banana"))

tally_delimited_string(df, fruits)

tally_delimited_string(df, fruits, count = TRUE)

tally_delimited_string(df, fruits, count = TRUE, names_repair = toupper)

}
